Altera banco de dados para MySQL e adiciona logica para encriptar id de no response

+++ src/Backend/MyRecipeBook.Infrastructure/Migrations/Versions/Version0000002.cs
@@ -12,22 +12,22 @@ public class Version0000002 : VersionBase
             .WithColumn("Title").AsString(255).NotNullable()
             .WithColumn("CookingTime").AsInt32().Nullable()
             .WithColumn("Difficulty").AsInt32().Nullable()
-            .WithColumn("UserId").AsInt32().NotNullable().ForeignKey("FK_Recipe_User_Id", "Users", "Id");
+            .WithColumn("UserId").AsInt64().NotNullable().ForeignKey("FK_Recipe_User_Id", "Users", "Id");
 
         CreateTable("Ingredients")
             .WithColumn("Item").AsString().NotNullable()
-            .WithColumn("RecipeId").AsInt32().NotNullable().ForeignKey("FK_Ingredient_RecipeId", "Recipes", "Id")
+            .WithColumn("RecipeId").AsInt64().NotNullable().ForeignKey("FK_Ingredient_RecipeId", "Recipes", "Id")
             .OnDelete(Rule.Cascade); // Delete cascade
         
         CreateTable("Instructions")
             .WithColumn("Step").AsInt32().NotNullable()
             .WithColumn("Text").AsString(2000).NotNullable()
-            .WithColumn("RecipeId").AsInt32().NotNullable().ForeignKey("FK_Instruction_RecipeId", "Recipes", "Id")
+            .WithColumn("RecipeId").AsInt64().NotNullable().ForeignKey("FK_Instruction_RecipeId", "Recipes", "Id")
             .OnDelete(Rule.Cascade); 
         
         CreateTable("DishTypes")
             .WithColumn("Type").AsInt32().NotNullable()
-            .WithColumn("RecipeId").AsInt32().NotNullable().ForeignKey("FK_DishType_RecipeId", "Recipes", "Id")
+            .WithColumn("RecipeId").AsInt64().NotNullable().ForeignKey("FK_DishType_RecipeId", "Recipes", "Id")
             .OnDelete(Rule.Cascade); 
     }
 }
